feat: opt-in fp16 loading for the image detector
PRISM_IMAGE_FP16 casts the CNN-ViT model + input to half precision,
roughly halving its resident memory (376MB -> ~188MB) — needed to fit
inside Render's 512MB free instance alongside torch/transformers
overhead. Also fixes GradCAM's heatmap output, which cv2.resize can't
handle in float16, by casting the (tiny) CAM array back to float32.
Off by default so local dev/training stays fp32.

+++ api/modules/image/cam.py
@@ -34,7 +34,7 @@ class GradCAM:
             return None
         weights = self._gradients.mean(dim=(2, 3), keepdim=True)   # (B, C, 1, 1)
         cam = (weights * self._activations).sum(dim=1, keepdim=True)  # (B, 1, H, W)
-        cam = torch.relu(cam).squeeze().cpu().numpy()                 # (H, W)
+        cam = torch.relu(cam).squeeze().cpu().float().numpy()         # (H, W) — cv2 has no float16 resize kernel
         lo, hi = cam.min(), cam.max()
         if hi > lo:
             cam = (cam - lo) / (hi - lo)
